requirements.txt
numpy
langdetect
googletrans==4.0.0-rc1
spacy
secure-smtplib
pandas
wordcloud
python-docx
PyPDF2
pdf2image
joblib
itertools
click==6.7
os-sys==2.1.4
Flask==1.0.2
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
six==1.11.0
Werkzeug==0.14.1
WTForms==2.2.1
nltk>=3.4.5
scikit-learn>=0.18
requests
